+++ internal/auth/event_handler_auth_user_projection.go
@@ -90,7 +90,6 @@ func (h *AuthUserProjectionEventHandler) handleUserRegistered(ctx context.Contex
 	if !ok {
 		return eventstore.ErrNotFound
 	}
-	println("1: ", resolved.Event.EventID)
 	return h.writer.CreateRegisteredUserAccount(ctx, RegisterUserResult{
 		EventID:      resolved.Event.EventID,
 		Email:        protectedpii.MustDecryptEventStringWithDataKey(protector, subjectKey, resolved.Event.Data, UserRegisteredEmailField),